Extract JobStatus class from DataImport
Connects to #12

This commit refactors the client by extracting a reusable JobStatus class from the DataImport class, letting each class have a single responsibility. This arguably improves / simplifies the interaction in the following ways (see the README for more):

* Avoid the "yo dawg" situation where the consumer has to ask twice to run the import
* Instead of returning the client response from `DataImport#import`, return the `JobStatus` instance
* Rename `JobStatus#job_status` to `JobStatus#status`
* Rename `JobStatus#wait` to `JobStatus#wait_until_complete` (align with DSC's similar functionality)
